chore: achieve zero-warning build with clippy fixes
- Fix all compiler warnings (dead_code annotations for API structs)
- Fix all clippy warnings (39 total):
  - Rename from_str() to parse() to avoid FromStr trait confusion
  - Simplify boolean expressions using is_none_or()
  - Use strip_prefix() instead of manual string slicing
  - Change &PathBuf to &Path in function signatures
  - Add Default impl for AuthBridge
  - Remove unnecessary type casts
  - Add missing transmute annotations
- Auto-fix various style improvements (map_or, closures, etc.)

+++ engine/crates/minna-graph/src/ring_engine.rs
@@ -298,7 +298,7 @@ impl RingEngine {
         let node_count = store.node_count().await?;
 
         // If graph has grown significantly, recalculate
-        if node_count as i64 > total_assigned * 2 {
+        if node_count > total_assigned * 2 {
             return Ok(true);
         }
